feat:用户查询商家收藏，正在写所有销量、用户查看消息

+++ src/main/java/com/example/pj/mapper/MessageMapper.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Mapper
 public interface MessageMapper {
     @Select("SELECT * FROM message WHERE user_id = #{userId}")
-    List<Message> getMessagesByUserId(int userId);
+    List<Message> getMessagesByUserId(Long userId);
 }
